Separate archival helpers from target-only helpers
