+++ src/main/java/io/cryostat/recordings/RecordingTargetHelper.java
@@ -37,17 +37,10 @@
  */
 package io.cryostat.recordings;
 
-import java.io.IOException;
-import java.io.InputStream;
-import java.net.URISyntaxException;
-import java.nio.file.Path;
-import java.time.temporal.ChronoUnit;
 import java.util.Map;
 import java.util.Optional;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
-
-import javax.inject.Named;
 
 import org.openjdk.jmc.common.unit.IConstrainedMap;
 import org.openjdk.jmc.flightrecorder.configuration.events.EventOptionID;
@@ -55,24 +48,18 @@ import org.openjdk.jmc.flightrecorder.configuration.recording.RecordingOptionsBu
 import org.openjdk.jmc.rjmx.services.jfr.IEventTypeInfo;
 import org.openjdk.jmc.rjmx.services.jfr.IRecordingDescriptor;
 
-import io.cryostat.MainModule;
 import io.cryostat.commands.internal.EventOptionsBuilder;
 import io.cryostat.core.net.JFRConnection;
-import io.cryostat.core.sys.Clock;
-import io.cryostat.core.sys.FileSystem;
 import io.cryostat.core.templates.TemplateType;
 import io.cryostat.messaging.notifications.NotificationFactory;
 import io.cryostat.net.ConnectionDescriptor;
 import io.cryostat.net.TargetConnectionManager;
-import io.cryostat.net.reports.ReportService;
 import io.cryostat.net.web.http.HttpMimeType;
-import io.cryostat.platform.PlatformClient;
-import io.cryostat.util.URIUtil;
 
 import org.apache.commons.lang3.StringUtils;
 import org.apache.commons.lang3.tuple.Pair;
 
-public class RecordingHelper {
+public class RecordingTargetHelper {
 
     private static final String NOTIFICATION_CATEGORY = "RecordingCreated";
 
@@ -82,28 +69,13 @@ public class RecordingHelper {
     private final TargetConnectionManager targetConnectionManager;
     private final EventOptionsBuilder.Factory eventOptionsBuilderFactory;
     private final NotificationFactory notificationFactory;
-    private final FileSystem fs;
-    private final Path recordingsPath;
-    private final Clock clock;
-    private final PlatformClient platformClient;
-    private final ReportService reportService;
 
-    RecordingHelper(
-            FileSystem fs,
-            @Named(MainModule.RECORDINGS_PATH) Path recordingsPath,
+    RecordingTargetHelper(
             TargetConnectionManager targetConnectionManager,
             EventOptionsBuilder.Factory eventOptionsBuilderFactory,
-            Clock clock,
-            PlatformClient platformClient,
-            ReportService reportService,
             NotificationFactory notificationFactory) {
-        this.fs = fs;
-        this.recordingsPath = recordingsPath;
         this.targetConnectionManager = targetConnectionManager;
         this.eventOptionsBuilderFactory = eventOptionsBuilderFactory;
-        this.clock = clock;
-        this.platformClient = platformClient;
-        this.reportService = reportService;
         this.notificationFactory = notificationFactory;
     }
 
@@ -142,45 +114,6 @@ public class RecordingHelper {
                             .build()
                             .send();
                     return desc;
-                });
-    }
-
-    public String saveRecording(ConnectionDescriptor connectionDescriptor, String recordingName)
-            throws Exception {
-
-        String saveName =
-                targetConnectionManager.executeConnectedTask(
-                        connectionDescriptor,
-                        connection -> {
-                            Optional<IRecordingDescriptor> descriptor =
-                                    getDescriptorByName(connection, recordingName);
-
-                            if (descriptor.isPresent()) {
-                                return writeRecordingToDestination(connection, descriptor.get());
-                            } else {
-                                throw new RecordingNotFoundException(recordingName);
-                            }
-                        });
-
-        return saveName;
-    }
-
-    public void deleteRecording(ConnectionDescriptor connectionDescriptor, String recordingName)
-            throws Exception {
-
-        targetConnectionManager.executeConnectedTask(
-                connectionDescriptor,
-                connection -> {
-                    Optional<IRecordingDescriptor> descriptor =
-                            getDescriptorByName(connection, recordingName);
-
-                    if (descriptor.isPresent()) {
-                        connection.getService().close(descriptor.get());
-                        reportService.delete(connectionDescriptor, recordingName);
-                    } else {
-                        throw new RecordingNotFoundException(recordingName);
-                    }
-                    return null;
                 });
     }
 
@@ -257,55 +190,5 @@ public class RecordingHelper {
         }
 
         return builder.build();
-    }
-
-    private String writeRecordingToDestination(
-            JFRConnection connection, IRecordingDescriptor descriptor) throws Exception {
-        String recordingName = descriptor.getName();
-        if (recordingName.endsWith(".jfr")) {
-            recordingName = recordingName.substring(0, recordingName.length() - 4);
-        }
-
-        // TODO: To avoid having to perform this lookup each time, we should implement
-        // something like a map from targetIds to corresponding ServiceRefs
-        String targetName =
-                platformClient.listDiscoverableServices().stream()
-                        .filter(
-                                serviceRef -> {
-                                    try {
-                                        return serviceRef
-                                                        .getServiceUri()
-                                                        .equals(
-                                                                URIUtil.convert(
-                                                                        connection.getJMXURL()))
-                                                && serviceRef.getAlias().isPresent();
-                                    } catch (URISyntaxException | IOException ioe) {
-                                        return false;
-                                    }
-                                })
-                        .map(s -> s.getAlias().get())
-                        .findFirst()
-                        .orElse(connection.getHost())
-                        .replaceAll("[\\._]+", "-");
-
-        String timestamp =
-                clock.now().truncatedTo(ChronoUnit.SECONDS).toString().replaceAll("[-:]+", "");
-        String destination = String.format("%s_%s_%s", targetName, recordingName, timestamp);
-        // TODO byte-sized rename limit is arbitrary. Probably plenty since recordings are also
-        // differentiated by second-resolution timestamp
-        byte count = 1;
-        while (fs.exists(recordingsPath.resolve(destination + ".jfr"))) {
-            destination =
-                    String.format("%s_%s_%s.%d", targetName, recordingName, timestamp, count++);
-            if (count == Byte.MAX_VALUE) {
-                throw new IOException(
-                        "Recording could not be savedFile already exists and rename attempts were exhausted.");
-            }
-        }
-        destination += ".jfr";
-        try (InputStream stream = connection.getService().openStream(descriptor, false)) {
-            fs.copy(stream, recordingsPath.resolve(destination));
-        }
-        return destination;
     }
 }
