+++ src/main/java/io/cryostat/rules/RuleProcessor.java
@@ -58,7 +58,8 @@ import io.cryostat.net.TargetConnectionManager;
 import io.cryostat.platform.PlatformClient;
 import io.cryostat.platform.ServiceRef;
 import io.cryostat.platform.TargetDiscoveryEvent;
-import io.cryostat.recordings.RecordingHelper;
+import io.cryostat.recordings.RecordingArchiveHelper;
+import io.cryostat.recordings.RecordingTargetHelper;
 import io.cryostat.rules.RuleRegistry.RuleEvent;
 import io.cryostat.util.events.Event;
 import io.cryostat.util.events.EventListener;
@@ -74,7 +75,8 @@ public class RuleProcessor
     private final CredentialsManager credentialsManager;
     private final RecordingOptionsBuilderFactory recordingOptionsBuilderFactory;
     private final TargetConnectionManager targetConnectionManager;
-    private final RecordingHelper recordingHelper;
+    private final RecordingArchiveHelper recordingArchiveHelper;
+    private final RecordingTargetHelper recordingTargetHelper;
     private final PeriodicArchiverFactory periodicArchiverFactory;
     private final Logger logger;
 
@@ -87,7 +89,8 @@ public class RuleProcessor
             CredentialsManager credentialsManager,
             RecordingOptionsBuilderFactory recordingOptionsBuilderFactory,
             TargetConnectionManager targetConnectionManager,
-            RecordingHelper recordingHelper,
+            RecordingArchiveHelper recordingArchiveHelper,
+            RecordingTargetHelper recordingTargetHelper,
             PeriodicArchiverFactory periodicArchiverFactory,
             Logger logger) {
         this.platformClient = platformClient;
@@ -96,7 +99,8 @@ public class RuleProcessor
         this.credentialsManager = credentialsManager;
         this.recordingOptionsBuilderFactory = recordingOptionsBuilderFactory;
         this.targetConnectionManager = targetConnectionManager;
-        this.recordingHelper = recordingHelper;
+        this.recordingArchiveHelper = recordingArchiveHelper;
+        this.recordingTargetHelper = recordingTargetHelper;
         this.periodicArchiverFactory = periodicArchiverFactory;
         this.logger = logger;
         this.tasks = new HashMap<>();
@@ -168,7 +172,7 @@ public class RuleProcessor
                                 serviceRef,
                                 credentialsManager,
                                 rule,
-                                recordingHelper,
+                                recordingArchiveHelper,
                                 this::archivalFailureHandler),
                         rule.getArchivalPeriodSeconds(),
                         rule.getArchivalPeriodSeconds(),
@@ -226,8 +230,9 @@ public class RuleProcessor
                         builder = builder.maxSize(rule.getMaxSizeBytes());
                     }
                     Pair<String, TemplateType> template =
-                            recordingHelper.parseEventSpecifierToTemplate(rule.getEventSpecifier());
-                    recordingHelper.startRecording(
+                            recordingTargetHelper.parseEventSpecifierToTemplate(
+                                    rule.getEventSpecifier());
+                    recordingTargetHelper.startRecording(
                             connectionDescriptor,
                             builder.build(),
                             template.getLeft(),
